rework mixins for better coverage, add `spawnType` predicate.  fixes #31 (#34)
This is a significant rework of the mixins that should fill a lot of gaps in terms of spawns that can be blocked.  For example, spawn blocking during chunk and structure generation is very airtight now.

It also adds the option to filter based on spawnType - see sample config for details.